Prototype of legacy configuration split
