+++ src/Glpi/Kernel/Kernel.php
@@ -73,6 +73,7 @@ final class Kernel extends BaseKernel
         $projectDir = $this->getProjectDir();
 
         $container->import($projectDir . '/dependency_injection/services.php', 'php');
+        $container->import($projectDir . '/dependency_injection/legacyConfigProviders.php', 'php');
     }
 
     protected function configureRoutes(RoutingConfigurator $routes): void
